pyproject.toml
[project]
name = "panoptes"
version = "0.1.0"
description = ""
authors = [
    {name = "Alessandro Monetti",email = "alessandromonetti@outlook.it"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "typeguard (>=4.4.2,<5.0.0)",
    "selenium (>=4.32.0,<5.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "shodan (>=1.31.0,<2.0.0)",
    "wappalyzer (>=1.0.13,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "rich-click (>=1.8.9,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "markdown2 (>=2.5.3,<3.0.0)",
    "weasyprint (>=65.1,<66.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]

[tool.poetry]
packages = [{include = "panoptes", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
